+++ CHANGELOG.md
@@ -8,8 +8,13 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - PHP 7.1 scalar type declarations and return types
 ### Removed
-- Removed support for PHP 5.x and 7.0 as they are no longer
+- **BC break**: Removed support for PHP 5.x and 7.0 as they are no longer
 [actively supported](https://php.net/supported-versions.php) by the PHP project
+- **BC break**: Removed deprecated methods:
+    - `Factory::fromFile()`
+    - `Factory::fromString()`
+    - `Factory::decodeHeader()`
+    - `Factory::parseEmailAddresses()`
 
 ## [2.1.1] - 2018-03-19
 ### Fixed